requirements.txt
asgiref==3.7.2
coverage==7.4.0
Django==5.0.1
djangorestframework==3.14.0
Markdown==3.5.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
